refactor: Move fetch_data module under src/utils/
